requirements.txt
bokeh==2.2.3
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.35.0
pandas==1.1.5